Refactor the orchpkg package
- Moved the logtools.go and configtools.go files to a new package fyrwatch/fyrmesh/tools in the root directory.
- Renamed the orchpkg package to orch.
- Refactored tool lib use within the orch package.
