+++ tools/configtools.go
@@ -1,4 +1,4 @@
-package orch
+package tools
 
 import (
 	"encoding/json"
